backend/requirements.txt
# — seus pacotes —
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
deepface==0.0.93
Pillow==9.5.0
python-multipart==0.0.5
numpy<2
python-dotenv
pytest
opencv-python
dlib-bin
requests>=2.27.1
tqdm>=4.30.0
gdown>=3.10.1
Flask>=1.1.2
flask-cors>=4.0.1
fire>=0.4.0
gunicorn>=20.1.0
pandas==2.3.2
mtcnn>=0.1.0
retina-face>=0.0.1
click>=7.0
h11>=0.8
dnspython<3.0.0,>=1.16.0
starlette==0.27.0
pydantic==1.10.22

# — pins de compat Windows + Py3.10 + TF 2.10 —
tensorflow==2.10.1
# TF 2.10.1 exige protobuf <3.20
protobuf==3.19.6
# Evita Keras 3.x; combina com TF 2.10.x
keras==2.10.0


